Rename: make folder and distribute codes
