+++ src/main/java/com/pettrip/app/controller/CourseController.java
@@ -1,10 +1,8 @@
 package com.pettrip.app.controller;
 import com.pettrip.app.dto.course.CourseDTO;
-import com.pettrip.service.CourseService;
+import com.pettrip.service.course.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/course")
