+++ src/main/java/com/pettrip/service/user/UserService.java
@@ -1,10 +1,7 @@
-package com.pettrip.service;
+package com.pettrip.service.user;
 
-import com.pettrip.domain.User;
-import com.pettrip.app.dto.JoinDTO;
 import com.pettrip.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
